Don't store separate ms data; create calendar folios from $.kmw; also set sc_cal_manifest correctly;
Previously was looking up ms from ms_id and storing in $.kui.manuscript and
using manuscripts widget code's $.kmw as well to store and update manuscripts
service with new data. For now, leaving all MS storage and updating to the
manuscripts widget. May want to make ms data storage and updates independent
of the ms widget in the future.